backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Image Processing
pillow==10.4.0
numpy>=1.24.0,<2.0.0

# Visualization & Reports
matplotlib>=3.8.0
scipy>=1.11.0
reportlab>=4.0.0

# Machine Learning (use tensorflow-cpu for smaller size on free tier)
tensorflow-cpu==2.15.0
